% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downlaod_images.R
\name{downlaod_images}
\alias{downlaod_images}
\title{Download iNaturalist images}
\usage{
downlaod_images(dat, size = "medium", outpath = ".")
}
\arguments{
\item{dat}{data frame containing taxonomic list}

\item{size}{image size. Values accepted are "small", "medium", "large", "original". Deafault(medium).}

\item{outpath}{path to the output folder}
}
\description{
Download iNaturalist images
}
\examples{
\dontrun{
downlaod_images(inatdata)
}
}
\seealso{
Other image functions: \code{\link{format_inat_images}}
}
\concept{image functions}
