% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_inat_images.R
\name{format_inat_images}
\alias{format_inat_images}
\title{Download formated iNaturalist images}
\usage{
format_inat_images(obs_id, outpath = ".", verbose = TRUE)
}
\arguments{
\item{obs_id}{iNaturalist id of the record}

\item{outpath}{path to the output folder}

\item{verbose}{print out comments while processing the images}
}
\description{
Download iNaturalist images for a recod and format them with metadata
}
\examples{
\dontrun{
format_inat_images(inatrecid)
}
}
\seealso{
Other image functions: \code{\link{downlaod_images}}
}
\concept{image functions}
